refactor: streamline Tailwind CSS configuration and remove unused components
- Removed extensive color definitions from tailwind.config.ts to simplify configuration.
- Deleted the StyleTest component as it was no longer needed.
- Updated page.tsx to enhance structure and styling by wrapping the title in a div with prose class.
- Refactored ThemeProvider to include getImplicitThemePreference function directly, improving theme management.
- Removed the now redundant getImplicitThemePreference utility file.
- Adjusted ThemeSelector styles for better visual consistency.
- Cleaned up global CSS by removing unused CSS variables and applying Tailwind utility classes for background and text colors.
