+++ src/lib/providers/Theme/ThemeSelector.tsx
@@ -37,7 +37,7 @@ export const ThemeSelector: React.FC = () => {
     <Select onValueChange={onThemeChange} value={value}>
       <SelectTrigger
         aria-label="Select a theme"
-        className="w-auto bg-transparent gap-2 pl-0 md:pl-3 border-none"
+        className="w-auto bg-transparent gap-2 pl-0 md:pl-3 border-radius-md dark:text-white dark:border-white"
       >
         <SelectValue placeholder="Theme" />
       </SelectTrigger>
