--- src/lib/utils/getImplicitThemePreference.ts
@@ -1,11 +1,0 @@
-export const getImplicitThemePreference = (): 'dark' | 'light' | null => {
-  const mediaQuery = '(prefers-color-scheme: dark)'
-  const mql = window.matchMedia(mediaQuery)
-  const hasImplicitPreference = typeof mql.matches === 'boolean'
-
-  if (hasImplicitPreference) {
-    return mql.matches ? 'dark' : 'light'
-  }
-
-  return null
-}
